FE-36 [FEAT] 워크스페이스에서 보드전체, 보드선택했을때 상세로 라우팅 설정
